Brian O'Conner and Roman Pearce (Tyrese Gibson) team up to go undercover to speed up docker builds in exchange for a smoother development workflow.

Build type     | Build Command | Time on `master` | Time on `build-opt`
----------------|----------------------|----------------------|-------------------------
Fresh clone | `./gradlew buildImage` | 2m 22s | 2m 1s |
Clean build  | `./gradlew clean && ./gradlew buildImage` | 1m 23s | 33s |
Rebuild (no changes) | `./gradlew buildImage` | 1m 18s | 31s |